+++ src/main/java/frc/robot/extras/sim/SimGyro.java
@@ -14,9 +14,9 @@ import edu.wpi.first.units.measure.LinearAcceleration;
 import edu.wpi.first.units.measure.LinearVelocity;
 import frc.robot.extras.sim.SimArena.SimEnvTiming;
 import frc.robot.extras.sim.configs.SimGyroConfig;
-import frc.robot.extras.sim.utils.RuntimeLog;
-import frc.robot.extras.sim.utils.mathutils.MeasureMath.XY;
-import frc.robot.extras.sim.utils.mathutils.SimCommonMath;
+import frc.robot.extras.util.utils.RuntimeLog;
+import frc.robot.extras.util.utils.mathutils.SimCommonMath;
+import frc.robot.extras.util.utils.mathutils.MeasureMath.XY;
 
 import java.util.function.BiConsumer;
