+++ src/main/java/frc/robot/extras/sim/SimWorld.java
@@ -9,8 +9,8 @@ import frc.robot.extras.sim.configs.SimMechanismConfig;
 import frc.robot.extras.sim.configs.SimSwerveConfig;
 import frc.robot.extras.sim.configs.SimSwerveModuleConfig;
 import frc.robot.extras.sim.configs.SimSwerveModuleConfig.WheelCof;
-import frc.robot.extras.sim.utils.DCMotorExt;
-import frc.robot.extras.sim.utils.GearRatio;
+import frc.robot.extras.util.utils.DCMotorExt;
+import frc.robot.extras.util.utils.GearRatio;
 import frc.robot.subsystems.swerve.SwerveConstants.DriveConstants;
 import frc.robot.subsystems.swerve.SwerveConstants.ModuleConstants;
 import static edu.wpi.first.units.Units.*;
