+++ src/main/java/frc/robot/extras/util/utils/FrcBody.java
@@ -1,4 +1,4 @@
-package frc.robot.extras.sim.utils;
+package frc.robot.extras.util.utils;
 
 import static edu.wpi.first.units.Units.NewtonMeters;
 import static edu.wpi.first.units.Units.RadiansPerSecond;
@@ -14,8 +14,8 @@ import edu.wpi.first.units.measure.Torque;
 import edu.wpi.first.util.struct.Struct;
 import edu.wpi.first.util.struct.StructGenerator;
 import edu.wpi.first.util.struct.StructSerializable;
-import frc.robot.extras.sim.utils.geometry.Velocity2d;
 import frc.robot.extras.util.GeomUtil;
+import frc.robot.extras.util.utils.geometry.Velocity2d;
 
 public class FrcBody extends Body {
     public record FrcBodySnapshot(
