+++ src/main/java/frc/robot/extras/util/utils/geometry/Velocity2d.java
@@ -1,4 +1,4 @@
-package frc.robot.extras.sim.utils.geometry;
+package frc.robot.extras.util.utils.geometry;
 
 import static edu.wpi.first.units.Units.MetersPerSecond;
