+++ src/main/java/frc/robot/extras/util/utils/geometry/Velocity3d.java
@@ -1,4 +1,4 @@
-package frc.robot.extras.sim.utils.geometry;
+package frc.robot.extras.util.utils.geometry;
 
 import java.util.Objects;
