+++ src/main/java/frc/robot/extras/util/utils/mathutils/MassMath.java
@@ -1,4 +1,6 @@
-package frc.robot.extras.sim.utils.mathutils;
+package frc.robot.extras.util.utils.mathutils;
+
+import static frc.robot.extras.util.utils.mathutils.MeasureMath.*;
 
 import edu.wpi.first.math.Pair;
 import edu.wpi.first.units.measure.AngularAcceleration;
@@ -8,7 +10,6 @@ import edu.wpi.first.units.measure.LinearAcceleration;
 import edu.wpi.first.units.measure.Mass;
 import edu.wpi.first.units.measure.MomentOfInertia;
 import edu.wpi.first.units.measure.Torque;
-import static frc.robot.extras.sim.utils.mathutils.MeasureMath.*;
 
 public class MassMath {
     public record PhysicsMass(Mass mass, MomentOfInertia moi) {
