+++ src/main/java/frc/robot/extras/util/utils/mathutils/SimCommonMath.java
@@ -1,4 +1,4 @@
-package frc.robot.extras.sim.utils.mathutils;
+package frc.robot.extras.util.utils.mathutils;
 
 import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.geometry.Translation2d;
